Input, Events and Everything: Tested

+++ tester/src/game.h
@@ -10,7 +10,7 @@
 
 typedef struct gameState
 {
-    double deltaTime;
+    float deltaTime;
 } gameState;
